app1/requirements.txt
streamlit
python-dotenv
google.generativeai